+++ Sources/BraintreeCore/Braintree-Version.h
@@ -1,1 +1,1 @@
-#define BRAINTREE_VERSION (@"5.5.0")
+#define BRAINTREE_VERSION (@"5.6.3")